+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change `renovate` configuration to only get upgrades for CAPA compatible versions of k8s dependencies.
 - Don't reconcile the `ShareReconciler` if only the status field has changed.
 
+### Fixed
+
+- Only remove `finalizer` after there are no more karpenter instances to terminate.
+
 ## [0.19.0] - 2025-06-17
 
 ### Added